+++ srcs/linear_algebra/matrices/matrix_utils.c
@@ -6,7 +6,7 @@
 /*   By: fnacarel <fnacarel@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/28 13:57:03 by fnacarel          #+#    #+#             */
-/*   Updated: 2023/08/03 14:56:09 by revieira         ###   ########.fr       */
+/*   Updated: 2023/08/11 19:06:22 by fnacarel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,9 +46,9 @@ int	is_invertible(t_matrix a)
 
 double	determinant(t_matrix a)
 {
-	int	i;
-	int	cof;
-	int	det;
+	int		i;
+	double	cof;
+	double	det;
 
 	i = -1;
 	det = 0;
